+++ src/routes/board.ts
@@ -1,9 +1,9 @@
-import { Router } from 'express';
+import { Router, Request, Response } from 'express';
 
-import * as boardService from '../services/board.js';
-import { BOARD_SCHEMA } from '../constants/validation.js';
-import { BOARD_ERRORS } from '../constants/errors.js';
-import { getValidationMiddleware } from '../middlewares/validation.js';
+import * as boardService from '../services/board';
+import { BOARD_SCHEMA } from '../constants/validation';
+import { BOARD_ERRORS } from '../constants/errors';
+import { getValidationMiddleware } from '../middlewares/validation';
 
 const validateBoard = getValidationMiddleware(BOARD_SCHEMA);
 
@@ -11,11 +11,11 @@ const router = Router();
 
 router
   .route('/')
-  .get(async (req, res) => {
+  .get(async (_req: Request, res: Response) => {
     const boards = await boardService.getAll();
     return res.json(boards);
   })
-  .post(validateBoard, async (req, res) => {
+  .post(validateBoard, async (req: Request, res: Response) => {
     const newBoard = await boardService.create(req.body);
 
     if (!newBoard) {
@@ -27,7 +27,7 @@ router
 
 router
   .route('/:id')
-  .get(async (req, res) => {
+  .get(async (req: Request<{ id: string }>, res: Response) => {
     const { id } = req.params;
     const board = await boardService.getById(id);
 
@@ -37,7 +37,7 @@ router
 
     return res.json(board);
   })
-  .put(validateBoard, async (req, res) => {
+  .put(validateBoard, async (req: Request<{ id: string }>, res: Response) => {
     const { id } = req.params;
     const updatedBoard = await boardService.updateById(id, req.body);
 
@@ -47,7 +47,7 @@ router
 
     return res.json(updatedBoard);
   })
-  .delete(async (req, res) => {
+  .delete(async (req: Request<{ id: string }>, res: Response) => {
     const { id } = req.params;
     const isBoardDeleted = await boardService.deleteById(id);
